Refactor search element
